requirements.txt
pyodbc>=4.0.0
mysql-connector-python>=8.0.0
pandas>=1.3.0
